move errors back to const
